SEMEX/requirements.txt
gitpython>=3.1
nbformat>=5.0